% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcopula.R
\docType{package}
\name{kcopula}
\alias{kcopula}
\title{kcopula: The Bivariate K-Copula}
\description{
kcopula provides the bivariate K-copula by Wollschläger and Schäfer (2016).
}
\details{
kcopula provides two functions:
    \itemize{
    \item \code{\link{pkcopula}} gives the distribution function of the
    bivariate K-copula.
    \item \code{\link{dkcopula}} gives the density of the bivariate K-copula.
    }
}
\examples{
## See README.md on GitHub for a comprehensive example.

}
\references{
Wollschläger, M. and Schäfer, R. (2016). Impact of nonstationarity on
    estimating and modeling empirical copulas of daily stock returns.
    \emph{Journal of Risk}, 19(1):1--23.
    \url{https://doi.org/10.21314/JOR.2016.342}.
    SSRN version: \url{https://ssrn.com/abstract=3533903}.

Chetalova, D., Wollschläger, M., and Schäfer, R. (2015).
    Dependence structure of market states.
    \emph{Journal of Statistical Mechanics: Theory and Experiment},
    2015(8):P08012. \url{https://doi.org/10.1088/1742-5468/2015/08/P08012}.
    SSRN version: \url{https://ssrn.com/abstract=3533951}.
}
\seealso{
Useful links:
    \itemize{
    \item \url{https://github.com/mlkremer/kcopula}
    \item Report bugs at \url{https://github.com/mlkremer/kcopula/issues}
    }
}
\author{
Marcel Kremer, \email{marcel.kremer@uni-due.de}
}
